fixint name typo in notification controller
